requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.52.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
tavily-python==0.5.0
rich==13.9.4
